Nettoyage
Petit nettoyage...

+++ srr/check.php
@@ -1,13 +1,13 @@
 <?php
     set_time_limit(0);
     error_reporting(E_ALL & ~E_NOTICE & ~E_DEPRECATED);
-    if(!isset($_GET['id'])){ die(); }
+    if(!isset($_GET['id']) or !is_numeric($_GET['id'])){ die(); }
     include 'include/simplepie.inc';
     $simple = new SimplePie();
     $simple->enable_cache(false);
     $simple->set_useragent('Mozilla/4.0 '.SIMPLEPIE_USERAGENT.' (with Simple RSS Reader)');
     $sqlite = new PDO('sqlite:include/data.db');
-    $sqlite->query('DELETE FROM items WHERE read=\'1\'');
+    $sqlite->query('DELETE FROM items WHERE read="1" AND user_id="' . $_GET['id'] . '"');
     $sqlite->query('VACUUM');
     $query = $sqlite->query('SELECT id,url FROM feeds WHERE user_id="' . $_GET['id'] . '"');
     while($response = $query->fetch()){
